+++ src/resources/views/pages/home.blade.php
@@ -63,6 +63,7 @@
                                 <li>Check <a href="https://github.com/redooor/redminportal/wiki/Private-API">list of Private API on github Wiki page</a></li>
                             </ul>
                         </li>
+                        <li>UI: Group permission management, able to edit group permission for any route. (issue 69)</li>
                     </ol>
                     <h4>Enhancements:</h4>
                     <ol>