+++ tests/middleware/DeleteAuthenticateTest.php
@@ -13,25 +13,26 @@ class DeleteAuthenticateTest extends BaseAuthenticateTest
             'admin/products/delete-variant-json/1',
         ];
         
+        // Redirected to previous page
         $this->test_redirects = [
-            'admin/announcements/delete/1'  => '/',
-            'admin/bundles/delete/1'        => '/',
-            'admin/categories/delete/1'     => '/',
-            'admin/coupons/delete/1'        => '/',
-            'admin/groups/delete/1'         => '/',
-            'admin/images/delete/1'         => '/',
-            'admin/mailinglists/delete/1'   => '/',
-            'admin/medias/delete/1'         => '/',
-            'admin/memberships/delete/1'    => '/',
-            'admin/modules/delete/1'        => '/',
-            'admin/orders/delete/1'         => '/',
-            'admin/pages/delete/1'          => '/',
-            'admin/portfolios/delete/1'     => '/',
-            'admin/posts/delete/1'          => '/',
-            'admin/products/delete/1'       => '/',
-            'admin/promotions/delete/1'     => '/',
-            'admin/purchases/delete/1'      => '/',
-            'admin/users/delete/1'          => '/',
+            'admin/announcements/delete/1'  => 'admin/products/delete-variant-json/1',
+            'admin/bundles/delete/1'        => 'admin/announcements/delete/1',
+            'admin/categories/delete/1'     => 'admin/bundles/delete/1',
+            'admin/coupons/delete/1'        => 'admin/categories/delete/1',
+            'admin/groups/delete/1'         => 'admin/coupons/delete/1',
+            'admin/images/delete/1'         => 'admin/groups/delete/1',
+            'admin/mailinglists/delete/1'   => 'admin/images/delete/1',
+            'admin/medias/delete/1'         => 'admin/mailinglists/delete/1',
+            'admin/memberships/delete/1'    => 'admin/medias/delete/1',
+            'admin/modules/delete/1'        => 'admin/memberships/delete/1',
+            'admin/orders/delete/1'         => 'admin/modules/delete/1',
+            'admin/pages/delete/1'          => 'admin/orders/delete/1',
+            'admin/portfolios/delete/1'     => 'admin/pages/delete/1',
+            'admin/posts/delete/1'          => 'admin/portfolios/delete/1',
+            'admin/products/delete/1'       => 'admin/posts/delete/1',
+            'admin/promotions/delete/1'     => 'admin/products/delete/1',
+            'admin/purchases/delete/1'      => 'admin/promotions/delete/1',
+            'admin/users/delete/1'          => 'admin/purchases/delete/1'
         ];
         
         $this->test_posts = [];
@@ -45,6 +46,7 @@ class DeleteAuthenticateTest extends BaseAuthenticateTest
         $this->test_pages = null; // Empty
         $this->test_pages = [];
         
+        // Redirected to previous page
         $this->test_redirects = null; // Empty
         $this->test_redirects = [
             'admin/bundles/delete/1'        => 'login/unauthorized',
@@ -60,11 +62,11 @@ class DeleteAuthenticateTest extends BaseAuthenticateTest
             'admin/products/delete/1'       => 'login/unauthorized',
             'admin/purchases/delete/1'      => 'login/unauthorized',
             'admin/users/delete/1'          => 'login/unauthorized',
-            'admin/announcements/delete/1'  => '/',
-            'admin/coupons/delete/1'        => '/',
-            'admin/pages/delete/1'          => '/',
-            'admin/posts/delete/1'          => '/',
-            'admin/promotions/delete/1'     => '/',
+            'admin/announcements/delete/1'  => 'admin/users/delete/1',
+            'admin/coupons/delete/1'        => 'admin/announcements/delete/1',
+            'admin/pages/delete/1'          => 'admin/coupons/delete/1',
+            'admin/posts/delete/1'          => 'admin/pages/delete/1',
+            'admin/promotions/delete/1'     => 'admin/posts/delete/1'
         ];
         
         $this->test_posts = null; // Empty
@@ -93,20 +95,21 @@ class DeleteAuthenticateTest extends BaseAuthenticateTest
         $this->test_pages = null; // Empty
         $this->test_pages = [];
         
+        // Redirected to previous page
         $this->test_redirects = null; // Empty
         $this->test_redirects = [
             'admin/bundles/delete/1'        => '/',
-            'admin/categories/delete/1'     => '/',
-            'admin/groups/delete/1'         => '/',
-            'admin/images/delete/1'         => '/',
-            'admin/mailinglists/delete/1'   => '/',
-            'admin/medias/delete/1'         => '/',
-            'admin/memberships/delete/1'    => '/',
-            'admin/modules/delete/1'        => '/',
-            'admin/orders/delete/1'         => '/',
-            'admin/portfolios/delete/1'     => '/',
-            'admin/products/delete/1'       => '/',
-            'admin/purchases/delete/1'      => '/',
+            'admin/categories/delete/1'     => 'admin/bundles/delete/1',
+            'admin/groups/delete/1'         => 'admin/categories/delete/1',
+            'admin/images/delete/1'         => 'admin/groups/delete/1',
+            'admin/mailinglists/delete/1'   => 'admin/images/delete/1',
+            'admin/medias/delete/1'         => 'admin/mailinglists/delete/1',
+            'admin/memberships/delete/1'    => 'admin/medias/delete/1',
+            'admin/modules/delete/1'        => 'admin/memberships/delete/1',
+            'admin/orders/delete/1'         => 'admin/modules/delete/1',
+            'admin/portfolios/delete/1'     => 'admin/orders/delete/1',
+            'admin/products/delete/1'       => 'admin/portfolios/delete/1',
+            'admin/purchases/delete/1'      => 'admin/products/delete/1',
             'admin/users/delete/1'          => 'login/unauthorized',
             'admin/announcements/delete/1'  => 'login/unauthorized',
             'admin/coupons/delete/1'        => 'login/unauthorized',